hent variabler som kun blir brukt i HjemlerService i HjemlerService
