+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/HjemlerService.kt
@@ -2,8 +2,9 @@ package no.nav.familie.ba.sak.kjerne.brev
 
 import no.nav.familie.ba.sak.integrasjoner.sanity.SanityService
 import no.nav.familie.ba.sak.kjerne.autovedtak.fødselshendelse.Resultat
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Målform
+import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
 import no.nav.familie.ba.sak.kjerne.vedtak.domene.VedtaksperiodeMedBegrunnelser
+import no.nav.familie.ba.sak.kjerne.vedtak.refusjonEøs.RefusjonEøsRepository
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
 import org.springframework.stereotype.Service
 
@@ -11,15 +12,18 @@ import org.springframework.stereotype.Service
 class HjemlerService(
     private val vilkårsvurderingService: VilkårsvurderingService,
     private val sanityService: SanityService,
+    private val persongrunnlagService: PersongrunnlagService,
+    private val refusjonEøsRepository: RefusjonEøsRepository,
 ) {
     fun hentHjemler(
         behandlingId: Long,
-        vedtaksperioder: List<VedtaksperiodeMedBegrunnelser>,
-        målform: Målform,
         vedtakKorrigertHjemmelSkalMedIBrev: Boolean = false,
-        refusjonEøsHjemmelSkalMedIBrev: Boolean,
-        erFritekstIBrev: Boolean,
+        sorterteVedtaksperioderMedBegrunnelser: List<VedtaksperiodeMedBegrunnelser>,
     ): String {
+        val refusjonEøsHjemmelSkalMedIBrev = refusjonEøsRepository.finnRefusjonEøsForBehandling(behandlingId)
+
+        val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(behandlingId = behandlingId)
+
         val vilkårsvurdering =
             vilkårsvurderingService.hentAktivForBehandling(behandlingId = behandlingId)
                 ?: error("Finner ikke vilkårsvurdering ved begrunning av vedtak")
@@ -28,14 +32,14 @@ class HjemlerService(
             vilkårsvurdering.finnOpplysningspliktVilkår()?.resultat == Resultat.IKKE_OPPFYLT
 
         return hentHjemmeltekst(
-            vedtaksperioder = vedtaksperioder,
+            vedtaksperioder = sorterteVedtaksperioderMedBegrunnelser,
             standardbegrunnelseTilSanityBegrunnelse = sanityService.hentSanityBegrunnelser(),
             eøsStandardbegrunnelseTilSanityBegrunnelse = sanityService.hentSanityEØSBegrunnelser(),
             opplysningspliktHjemlerSkalMedIBrev = opplysningspliktHjemlerSkalMedIBrev,
-            målform = målform,
+            målform = personopplysningGrunnlag.søker.målform,
             vedtakKorrigertHjemmelSkalMedIBrev = vedtakKorrigertHjemmelSkalMedIBrev,
-            refusjonEøsHjemmelSkalMedIBrev = refusjonEøsHjemmelSkalMedIBrev,
-            erFritekstIBrev = erFritekstIBrev,
+            refusjonEøsHjemmelSkalMedIBrev = refusjonEøsHjemmelSkalMedIBrev.isNotEmpty(),
+            erFritekstIBrev = sorterteVedtaksperioderMedBegrunnelser.any { it.fritekster.isNotEmpty() },
         )
     }
 }
